+++ pkg/dataplane/http/types.go
@@ -6,5 +6,5 @@ type NewContextInput struct {
 	HTTPClient     *fasthttp.Client
 	NumWorkers     int
 	RequestChanLen int
-	MaxConns       int64
+	MaxConns       int
 }
